[#3658] docs(hadoop-catalog): Add document about user authentication for Hadoop catalog (#3659)
### What changes were proposed in this pull request?

Add document about user authentication for Hadoop catalog

### Why are the changes needed?

To improve user experience.

Fix: #3658 

### Does this PR introduce _any_ user-facing change?

N/A

### How was this patch tested?

N/A.
