+++ catalogs/catalog-hadoop/src/main/java/com/datastrato/gravitino/catalog/hadoop/kerberos/KerberosConfig.java
@@ -23,14 +23,11 @@ public class KerberosConfig extends AuthenticationConfig {
   public static final String FETCH_TIMEOUT_SEC_KEY =
       "authentication.kerberos.keytab-fetch-timeout-sec";
 
-  public static final String IMPERSONATION_ENABLE_KEY =
-      "authentication.kerberos.impersonation-enable";
-
   public static final boolean DEFAULT_IMPERSONATION_ENABLE = false;
 
   public static final ConfigEntry<String> PRINCIPAL_ENTRY =
       new ConfigBuilder(PRINCIPAL_KEY)
-          .doc("The principal of the kerberos connection")
+          .doc("The principal of the Kerberos connection")
           .version(ConfigConstants.VERSION_0_5_1)
           .stringConf()
           .checkValue(StringUtils::isNotBlank, ConfigConstants.NOT_BLANK_ERROR_MSG)
@@ -38,7 +35,7 @@ public class KerberosConfig extends AuthenticationConfig {
 
   public static final ConfigEntry<String> KEYTAB_ENTRY =
       new ConfigBuilder(KEY_TAB_URI_KEY)
-          .doc("The keytab of the kerberos connection")
+          .doc("The keytab of the Kerberos connection")
           .version(ConfigConstants.VERSION_0_5_1)
           .stringConf()
           .checkValue(StringUtils::isNotBlank, ConfigConstants.NOT_BLANK_ERROR_MSG)
@@ -46,7 +43,7 @@ public class KerberosConfig extends AuthenticationConfig {
 
   public static final ConfigEntry<Integer> CHECK_INTERVAL_SEC_ENTRY =
       new ConfigBuilder(CHECK_INTERVAL_SEC_KEY)
-          .doc("The check interval of the kerberos connection for Hadoop catalog")
+          .doc("The check interval of the Kerberos connection for Hadoop catalog")
           .version(ConfigConstants.VERSION_0_5_1)
           .intConf()
           .checkValue(value -> value > 0, ConfigConstants.POSITIVE_NUMBER_ERROR_MSG)
@@ -54,26 +51,15 @@ public class KerberosConfig extends AuthenticationConfig {
 
   public static final ConfigEntry<Integer> FETCH_TIMEOUT_SEC_ENTRY =
       new ConfigBuilder(FETCH_TIMEOUT_SEC_KEY)
-          .doc("The fetch timeout of the kerberos connection")
+          .doc("The fetch timeout of the Kerberos connection")
           .version(ConfigConstants.VERSION_0_5_1)
           .intConf()
           .checkValue(value -> value > 0, ConfigConstants.POSITIVE_NUMBER_ERROR_MSG)
           .createWithDefault(2);
 
-  public static final ConfigEntry<Boolean> ENABLE_IMPERSONATION_ENTRY =
-      new ConfigBuilder(IMPERSONATION_ENABLE_KEY)
-          .doc("Whether to enable impersonation for the Hadoop catalog")
-          .version(ConfigConstants.VERSION_0_5_1)
-          .booleanConf()
-          .createWithDefault(DEFAULT_IMPERSONATION_ENABLE);
-
   public KerberosConfig(Map<String, String> properties) {
     super(properties);
     loadFromMap(properties, k -> true);
-  }
-
-  public boolean isImpersonationEnabled() {
-    return get(ENABLE_IMPERSONATION_ENTRY);
   }
 
   public String getPrincipalName() {
@@ -94,16 +80,6 @@ public class KerberosConfig extends AuthenticationConfig {
 
   public static final Map<String, PropertyEntry<?>> KERBEROS_PROPERTY_ENTRIES =
       new ImmutableMap.Builder<String, PropertyEntry<?>>()
-          .put(
-              IMPERSONATION_ENABLE_KEY,
-              PropertyEntry.booleanPropertyEntry(
-                  IMPERSONATION_ENABLE_KEY,
-                  "Whether to enable impersonation for the Hadoop catalog",
-                  false,
-                  true,
-                  DEFAULT_IMPERSONATION_ENABLE,
-                  false,
-                  false))
           .put(
               KEY_TAB_URI_KEY,
               PropertyEntry.stringImmutablePropertyEntry(
